add Alert popup on auth errors
