+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Artem", "beatfilm", "expolrer", "nomoredomains"]
+  "cSpell.words": ["Artem", "beatfilm", "expolrer", "hocs", "nomoredomains"]
 }
