+++ src/components/Alert/Alert.jsx
@@ -1,0 +1,11 @@
+import './Alert.css';
+
+function Alert({ isActiveAlert, messageAlert }) {
+  return (
+    <div className={isActiveAlert ? 'alert alert_active' : 'alert'}>
+      <p className="alert__text">{messageAlert}</p>
+    </div>
+  );
+}
+
+export default Alert;
